Move internel likelihood function from utils.py to likelihood.py.
